Added loadbalancer to charm

+++ src/tools/CollectorCharm.hpp
@@ -44,7 +44,7 @@ public:
         if(all){
             for (int i = 0; i < 4; i++) {
                 if(i==CTR_WALL)
-                    std::cout<<  "recv measure_ctr "<<arr[i+5] << std::endl;
+                    std::cout<<  "recv measure_ctr "<<(long)arr[i+5] << std::endl;
                 instance.measure_ctrs[i] = std::chrono::steady_clock::time_point(std::chrono::milliseconds((long)arr[i+5]));
 
             }
